+++ statsmile/data/__init__.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
-from .user import get_steam_user
-from .converter import converter
+from .getuser import get_steam_user
 from .getmatches import getting_matches_id
-from .update import update_matches_id
 from .getmatch import update_matches
+from .update import update_matches_id