+++ statsmile/handlers/__init__.py
@@ -10,4 +10,3 @@ from .players import PlayersHandler
 from .start import StartHandler
 from .match import MatchHandler
 from .events import EventsHandler
-from .heroes import HeroesHandler