+++ statsmile/handlers/auth.py
@@ -23,11 +23,11 @@ class AuthHandler(BaseHandler, tornado.auth.OpenIdMixin):
             steamid = claimed_id["claimed_id"][-17:]
             rv = self.application.db["users"].find_one({"steamid": steamid})
             if not rv:
-                user = get_steam_user(self.application.db, self.application.logger, steamid)
+                user = get_steam_user(self.application.db, steamid)
                 self.application.db["users"].insert(user)
                 self.set_secure_cookie("statsmile_user", json_encode(str(user["_id"])))
                 self.redirect("/user/{}".format(user["_id"]))
-                yield getting_matches_id(self.application.db, self.application.logger, steamid)
+                yield getting_matches_id(self.application.db, steamid)
             else:
                 self.set_secure_cookie("statsmile_user", json_encode(str(rv["_id"])))
                 self.redirect("/user/{}".format(rv["_id"]))